[ripd] bug #293: routemap set metric doesn't check for underflow correctly
2006-09-11 Paul Jakma <paul.jakma@sun.com>

	* rip_routemap.c: (route_set_metric) underflow check needs to
	  use signed, problem identified and diagnosed by Pavel
	  Nikiforov in bug #293.

+++ ripd/rip_routemap.c
@@ -530,7 +530,7 @@ route_set_metric (void *rule, struct prefix *prefix,
       else if (mod->type == metric_absolute)
 	rinfo->metric_out = mod->metric;
 
-      if (rinfo->metric_out < 1)
+      if ((signed int)rinfo->metric_out < 1)
 	rinfo->metric_out = 1;
       if (rinfo->metric_out > RIP_METRIC_INFINITY)
 	rinfo->metric_out = RIP_METRIC_INFINITY;
